backend/requirements.txt
cryptography==46.0.5
fastapi==0.135.1
httpx==0.28.1
ldap3==2.9.1
neo4j==6.1.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-sdk==1.38.0
paramiko==4.0.0
psycopg[binary]==3.2.12
pyotp==2.9.0
pydantic==2.11.7
pytest==8.4.2
redis==7.1.0
reportlab==4.4.3
starlette==0.49.1
uvicorn[standard]==0.35.0
xlsxwriter==3.2.8
